Develop ReactorManager module

+++ src/reactor_manager.h
@@ -40,7 +40,7 @@ class ServerReactorManager : NonCopyableMovable {
   typedef std::vector<EventManager*> EventManagers;
 
   ServerReactorManager(const NetAddress& listen_address,
-                       int io_thread_amount = 6,
+                       size_t io_thread_amount = 6,
                        bool should_reuse_port = false);
   ~ServerReactorManager();
